refactor(CardGame): 클래스 분리
-변경: CardGame 클래스를 생성을 통해 Application 내 게임 실행 로직 분리
